add more descriptive comment and move compass manager
