+++ common/src/main/java/com/wynntils/mc/utils/CompassManager.java
@@ -2,9 +2,8 @@
  * Copyright © Wynntils 2022.
  * This file is released under AGPLv3. See LICENSE for full license details.
  */
-package com.wynntils.managers;
+package com.wynntils.mc.utils;
 
-import com.wynntils.mc.utils.McUtils;
 import com.wynntils.utils.objects.Location;
 
 // TODO: Draw compass beam
